requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-slugify==4.0.1
SQLAlchemy==1.3.18
text-unidecode==1.3
Werkzeug==1.0.1